chore(shades):moved PartialElement to Models
