+++ packages/shades/src/models/index.ts
@@ -1,3 +1,4 @@
 export * from './children-list'
+export * from './partial-element'
 export * from './render-options'
 export * from './shade-component'
