+++ packages/shades/src/models/partial-element.ts
@@ -1,0 +1,3 @@
+export type PartialElement<T> = {
+  [K in keyof T]?: T[K] extends (((...args: any[]) => any) | null | undefined) ? T[K] : PartialElement<T[K]>
+}
